refactor suggestion experiment code to its own file
